Addressing code review

+++ aslam_cv/src/camera-pinhole.cc
@@ -31,7 +31,7 @@ PinholeCamera::PinholeCamera()
 PinholeCamera::PinholeCamera(const Eigen::VectorXd& intrinsics,
               uint32_t imageWidth, uint32_t imageHeight,
               aslam::Distortion::Ptr distortion) {
-  CHECK_EQ(kNumOfParams, intrinsics.size());
+  CHECK_EQ(static_cast<int>(kNumOfParams), intrinsics.size());
   CHECK_NOTNULL(distortion.get());
   setParameters(intrinsics);
   _distortion = distortion;
